🏗️ lifting props and teasing apart reusable components
