+++ .changeset/mighty-ears-breathe.md
@@ -1,0 +1,6 @@
+---
+'@myst-theme/site': patch
+'@myst-theme/book': patch
+---
+
+Renamed `Navigation` component and split for re-use in different (composed/multi-site) themes
