+++ packages/site/src/components/Navigation/index.tsx
@@ -1,7 +1,8 @@
 export { ThemeButton } from './ThemeButton.js';
 export { TopNav, NavItems, NavItem, DEFAULT_NAV_HEIGHT } from './TopNav.js';
-export { Navigation } from './Navigation.js';
-export { PrimarySidebar, InlineTableOfContents, useSidebarHeight } from './PrimarySidebar.js';
+export { Navigation, MobileNavigation, ConfigurableMobileNavigation } from './Navigation.js';
+export { PrimarySidebar, useSidebarHeight } from './PrimarySidebar.js';
+export { InlineTableOfContents } from './InlineTableOfContents.js';
 export { LoadingBar } from './Loading.js';
 export { ActionMenu } from './ActionMenu.js';
 export { HomeLink } from './HomeLink.js';
